requirements.txt
flask==3.0.3
xmltodict==0.13.0
requests==2.31.0
python-dotenv==1.0.1
google-genai==1.19.0
functions_framework==3.9.2
